Move toggable fieldsets to global listeners
